changed step for asking for password to decrypt wallet and added password check for retry
